early_video_app: add NV12 UBWC format support
1.change default decode output and render format from NV12 to NV12 UBWC.
2.add render frame rate control.

+++ early_video_app/inc/V4l2Decoder.h
@@ -37,7 +37,7 @@ public:
 	int setFrameRate(unsigned int numer, unsigned int denom);
 	int setDSResolution(unsigned int width, unsigned int height);
 	int getOperatingRate();
-	int getFrameRate();
+	float getFrameRate();
 	int drain();
 	int resume();
 	struct v4l2_format* getOutputFormat();
